Fix transitions
Wierd Blender thing: a cross dissolve consists of and input_1 and input_2 clip, but they are in the order they are selected! So selecting right handle before left handle and then add a cross transition would export wrong. Doing the opposite would go well. Strange. And all the info on the inputs were not about the parts of the clips used, but what came before or after the cross transition.